refactor / code abstraction
